move default and user config to own files
