+++ config/user_config.go
@@ -1,0 +1,83 @@
+package config
+
+import (
+	"fmt"
+	"gopkg.in/yaml.v3"
+	"strconv"
+)
+
+type UserConfig struct {
+	ImageColPaths []string `yaml:"image_col_paths"`
+	Interval      int      `yaml:"interval"`
+	Profile       string   `yaml:"profile"`
+	Alignment     string   `yaml:"default_alignment"`
+	Stretch       string   `yaml:"default_stretch"`
+	Opacity       float64  `yaml:"default_opacity"`
+}
+
+func (c *UserConfig) IsDefaultConfig() bool {
+	return false
+}
+
+func (c *UserConfig) IsUserConfig() bool {
+	return true
+}
+
+func (c *UserConfig) Unmarshal(data []byte) error {
+	err := yaml.Unmarshal(data, c)
+	if err != nil {
+		return fmt.Errorf("Failed to unmarshal user config: %s", err)
+	}
+	return nil
+}
+
+func (c *UserConfig) Log(configPath string) Config {
+	fmt.Println("------------------------------------------------------------------------------------")
+	fmt.Println("|", configPath)
+	fmt.Println("------------------------------------------------------------------------------------")
+	fmt.Println("| image_col_paths:")
+	for _, path := range c.ImageColPaths {
+		fmt.Printf("%-25s%s\n", "|", path)
+	}
+	fmt.Printf("|\n%-25s%s\n", "| interval:", strconv.Itoa(c.Interval))
+	fmt.Printf("%-25s%s\n", "| profile:", c.Profile)
+	fmt.Printf("%-25s%s\n", "| default_alignment:", c.Alignment)
+	fmt.Printf("%-25s%s\n", "| default_stretch:", c.Stretch)
+	fmt.Printf("%-25s%s\n", "| default_opacity:", strconv.FormatFloat(c.Opacity, 'f', -1, 64))
+	fmt.Println("------------------------------------------------------------------------------------")
+
+	return c
+}
+
+func (c *UserConfig) LogEdited(editedPaths map[string]string) Config {
+	fmt.Println("| edited: ")
+	fmt.Println("------------------------------------------------------------------------------------")
+	if _, ok := editedPaths["no changes made"]; ok {
+		fmt.Println("| no changes made")
+	} else {
+		for old, edited := range editedPaths {
+			if old != edited {
+				fmt.Printf("%-25s%s\n", "| old:", old)
+				fmt.Printf("%-25s%s\n", "| new:", edited)
+				fmt.Println("|")
+			}
+		}
+	}
+	fmt.Println("------------------------------------------------------------------------------------")
+
+	return c
+}
+
+func (c *UserConfig) LogRemoved(path map[string]struct{}) Config {
+	fmt.Print("| removed: ")
+	if _, ok := path["no changes made"]; ok {
+		fmt.Println("no changes made")
+	} else {
+		for path := range path {
+			fmt.Println(path)
+		}
+	}
+	fmt.Println("------------------------------------------------------------------------------------")
+
+	return c
+}
